ajuste das categorias com tipo invertido nos imports; refatoraccao e arquitetura de codigo
