+++ app/templates/board/pages/credit_card_dashboard_modal_upload_csv.html
@@ -33,13 +33,13 @@
                         <div class='col-md'>
                             <div class='card-body'>
                                 <div class='col-sm'>
-                                    <button class="btn btn-primary w-100 text-start mt-2" type="button"
+                                    <button class="btn btn-warning w-100 text-start mt-2" type="button"
                                             data-bs-toggle="collapse"
                                             data-bs-target="#collapseUploadNotes" aria-expanded="false"
                                             aria-controls="collapseUploadNotes">
                                         <div class='row mb'>
                                             <div class='col-sm'>
-                                                <h4 class='card-title text-center text-white'>
+                                                <h4 class='card-title text-center text-black'>
                                                     Observações sobre a importação do arquivo CSV
                                                 </h4>
                                             </div>
